Upload initial plugin files
The Snippet Genius WordPress plugin's initial set of files is added with this commit. CSS stylesheet and the main plugin file are all included. The foundation for handling unique code snippets in different areas of a WordPress website is provided by these files.

The plugin is licensed under GPL-3.0-or-later.

# New Features:
- Added main plugin file (snippet-genius-main.php)
- Included CSS stylesheet for styling the plugin interface

# License:
GPL-3.0-or-later